add Header and InputField components, restructure folder for pages
